Explore reduced configuration extension.
Do not register a MappingContext through the extension, instead expose a MappingContext through KeyValueAdapter so implementations can bring their own mapping context. They can also decide where to obtain the mapping context from and whether to expose it as bean at all. This is useful for a minimal configuration and removes the need to detect whether there is already a mapping context available.

Detection of mapping context (or even converters) can go either into the adapter or somewhere else as detecting bean registrations depends on configuration ordering. Any later bean registrations might be not visible to our extension and so we had always the drawback of requiring any customizations being colocated within the config class that uses the Enable…Repositories annotation.